Moved e2e test code out of TOSA backend
 * Unit tests now small and focused on Tosa flow
 * All output into tempfile locations
 * fixed use of capture configs
